new stuff for cli
